Rename to gtm slug because of ad-blockers. Reset dirName on toggle of app
