+++ packages/app-store/apps.keys-schemas.generated.ts
@@ -6,8 +6,8 @@ import { appKeysSchema as dailyvideo_zod_ts } from "./dailyvideo/zod";
 import { appKeysSchema as fathom_zod_ts } from "./fathom/zod";
 import { appKeysSchema as ga4_zod_ts } from "./ga4/zod";
 import { appKeysSchema as giphy_zod_ts } from "./giphy/zod";
-import { appKeysSchema as google_tag_manager_zod_ts } from "./google-tag-manager/zod";
 import { appKeysSchema as googlecalendar_zod_ts } from "./googlecalendar/zod";
+import { appKeysSchema as gtm_zod_ts } from "./gtm/zod";
 import { appKeysSchema as hubspot_zod_ts } from "./hubspot/zod";
 import { appKeysSchema as larkcalendar_zod_ts } from "./larkcalendar/zod";
 import { appKeysSchema as office365calendar_zod_ts } from "./office365calendar/zod";
@@ -31,8 +31,8 @@ export const appKeysSchemas = {
   fathom: fathom_zod_ts,
   ga4: ga4_zod_ts,
   giphy: giphy_zod_ts,
-  "google-tag-manager": google_tag_manager_zod_ts,
   googlecalendar: googlecalendar_zod_ts,
+  gtm: gtm_zod_ts,
   hubspot: hubspot_zod_ts,
   larkcalendar: larkcalendar_zod_ts,
   office365calendar: office365calendar_zod_ts,
