+++ packages/app-store/gtm/package.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://json.schemastore.org/package.json",
   "private": true,
-  "name": "@calcom/google-tag-manager",
+  "name": "@calcom/gtm",
   "version": "0.0.0",
   "main": "./index.ts",
   "dependencies": {
