+++ packages/app-store/sylapsvideo/DESCRIPTION.md
@@ -1,6 +1,6 @@
 ---
 items:
-  - /api/app-store/sylapsvideo/1.jpg
+  - 1.jpg
 ---
 
 Free Audio and Video Conferencing, Online Collaboration, Screen Sharing on web browser, mobile and desktop.
