+++ packages/app-store/sylapsvideo/config.json
@@ -2,10 +2,10 @@
   "/*": "Don't modify slug - If required, do it using cli edit command",
   "name": "Sylaps",
   "title": "Sylaps",
-  "slug": "sylaps",
+  "slug": "sylapsvideo",
   "type": "sylaps_video",
-  "imageSrc": "/api/app-store/sylapsvideo/icon.svg",
-  "logo": "/api/app-store/sylapsvideo/icon.svg",
+  "imageSrc": "icon.svg",
+  "logo": "icon.svg",
   "url": "https://cal.com/apps/sylaps",
   "variant": "conferencing",
   "categories": ["video"],
